+++ packages/sapa/src/App.js
@@ -33,8 +33,20 @@ export const start = (ElementClass, opt = {}) => {
     // id 유지
     app.id = $targetElement.el[COMPONENT_INSTANCE].id;
     app.render();
+    console.log(
+      "update render",
+      app.$el.el,
+      app.id,
+      app.$el.el[COMPONENT_INSTANCE].id
+    );
   } else {
     app.render($container);
+    console.log(
+      "render",
+      app.$el.el,
+      app.id,
+      app.$el.el[COMPONENT_INSTANCE].id
+    );
   }
   registRootElementInstance(app, $container);
 
@@ -60,7 +72,7 @@ export const hydrate = (ElementClass, opt = {}) => {
   const $targetElement = $container.firstChild;
 
   if ($targetElement) {
-    app.$el = Dom.create($targetElement.el);
+    app.$el = $targetElement;
     app.render();
   } else {
     app.render($container);